Adding checkpoint for laplace_eqn and heat_eqn submodules, both are partially implemented at an early testing phase; simulation grid made, initial conditions set using hardcoded function, Dirichlet boundary conditions set (Neumann and Periodic will be implemeneted in future), and filled contour image of initial state is rendered on submodule window. Currently, laplace_eqn can update solution image with 'Next' button, using a simple Jacobi iteration step method; heat_eqn has a disabled 'Next' button. More to come on PDEs modules. Made minor change to base_window [BaseWindow] to not print removed tmpfile for Windows OS and to check if a tmpfile was made at instantiation before attempting to remove said tmpfile, prevents crash for windows that don't use tmpfiles. Also made minor changes to FFTs submodules, removing unnecessary import statements.

+++ iea/utils/base_window.py
@@ -40,8 +40,8 @@ class BaseWindow:
   def __init__(self, configfile=None, launch=False, makeTempFile=False):
 
     self._arch = self._determine_arch()
-
-    if makeTempFile:
+    self._tmp = makeTempFile
+    if self._tmp:
       tempdict = {"suffix": ".png"}
       if self._arch == "windows":
         tempdict['delete'] = False
@@ -54,9 +54,9 @@ class BaseWindow:
 
 
   def __del__(self):
-    if self._arch == "windows":
+    if self._arch == "windows" and self._tmp:
       self._fout.close()
-      print(f"Removing Tempfile {self._fout.name}")
+      # print(f"Removing Tempfile {self._fout.name}")
       os.remove(self._fout.name)
